+++ akka-stream/src/main/scala/akka/stream/impl/ReactiveStreamsCompliance.scala
@@ -1,13 +1,12 @@
-/**
- * Copyright (C) 2014 Typesafe Inc. <http://www.typesafe.com>
- */
-package akka.stream
-
-import org.reactivestreams.{ Subscription, Subscriber, Publisher }
+package akka.stream.impl
 
 import scala.util.control.NonFatal
+import org.reactivestreams.{ Subscriber, Publisher, Subscription }
 
-object ReactiveStreamsConstants {
+/**
+ * INTERNAL API
+ */
+private[stream] object ReactiveStreamsCompliance {
 
   final val CanNotSubscribeTheSameSubscriberMultipleTimes =
     "can not subscribe the same subscriber multiple times (see reactive-streams specification, rules 1.10 and 2.12)"
